Tribute Page
A tribute page is basically an overview of someone whom we admire in our life. In this article, I will be creating a tribute webpage of Heavy Weight Champion Muhammad Alli using HTML and CSS. I will add an image of him at the center (below the title) and create a box beneath that image.